Fix: Solve bug compile

+++ Orchextra/Common/Interactors/ValidatorTrigger/ORCValidatorActionInterator.m
@@ -251,8 +251,8 @@ int ERROR_ACTION_NOT_FOUND = 5001;
         if (type && value)
         {
             if (response.action != nil && response.action.type != nil && response.action.urlString != nil && response.action.scheduleTime) {
-                [[ORCLog sharedInstance] logDebug: [NSString stringWithFormat: @"---- FOUND ACTION ---- \n ------> Trigger: %@, Value: %@\n ------> Action: %@, url: %@, Schedule: %d\n",
-                                                    type, value, response.action.type, response.action.urlString, response.action.scheduleTime]];
+                [[ORCLog sharedInstance] logDebug: [NSString stringWithFormat: @"---- FOUND ACTION ---- \n ------> Trigger: %@, Value: %@\n ------> Action: %@, url: %@, Schedule: %ld\n",
+                                                    type, value, response.action.type, response.action.urlString, (long)response.action.scheduleTime]];
             }
             else
             {
